+++ lambdas/package/index.py
@@ -4,12 +4,14 @@ disk and RAM pressure.
 
 Lambda functions can have up to 3GB of RAM and only 512MB of disk.
 """
+
+import io
+
 import boto3
+import pandas as pd
 
 from t4_lambda_shared.decorator import api, validate
 from t4_lambda_shared.utils import get_default_origins, make_json_response
-
-from .package_browse import get_logical_key_folder_view, load_df
 
 S3_DOMAIN_SUFFIX = '.amazonaws.com'
 
@@ -35,6 +37,42 @@ SCHEMA = {
     'required': ['bucket', 'key', 'access_key', 'secret_key'],
     'additionalProperties': False
 }
+
+
+
+def load_df(s3response):
+    """
+    Read a streaming response from s3 select into a
+    Pandas DataFrame
+    """
+    buffer = io.StringIO()
+    for event in s3response['Payload']:
+        if 'Records' in event:
+            records = event['Records']['Payload'].decode('utf-8')
+            buffer.write(records)
+        elif 'Stats' in event:
+            statsDetails = event['Stats']['Details']
+    buffer.seek(0)
+    df = pd.read_json(buffer, lines=True)
+    return df, statsDetails
+    
+def get_logical_key_folder_view(df, prefix=None):
+    """
+    Post process a set of logical keys to return only the
+    top-level folder view (a special case of the s3-select
+    lambda).
+    """
+    if prefix:
+        col = df.logical_key.str.slice(start=len(prefix))
+    else:
+        col = df.logical_key
+        
+    # matches all strings; everything before and including the first
+    # / is extracted
+    folder = col.dropna().str.extract('([^/]+/?).*')[0].unique().tolist()
+    return folder
+ 
+
 
 @api(cors_origins=get_default_origins())
 @validate(SCHEMA)
